Prevent buffer overflows with buf with read_line()
In main(), buf will read from unixfd with read_line() function to read byte by byte
rather than try to copy the entire message into the buffer all at once. This will prevent
buffer overflows on buf. As a consqeunce of this, all messages sent to irccd must end with '\n'
or else the server will continue to block on socket reading. This might want to be looked at.